results for paper draft

+++ optimizations/dsc/test_longTrace/bnb_test_gas.js
@@ -6,8 +6,8 @@ const {
   time, // Assertions for transactions that should fail
 } = require('@openzeppelin/test-helpers');
 
-const helper = require("./helper_functions");
-// const helper = require("./helper_functions_geth");
+// const helper = require("./helper_functions");
+const helper = require("./helper_functions_geth");
 const fs = require('fs');
 const path = require('path');
 const lowerBoundInput = 10;
@@ -83,10 +83,13 @@ helper.range(transactionCounts).forEach(l => {
       let approveAmount_1 = helper.random(1, (transferAmount+1)/2);
       let approveAmount_2 = helper.random(1, (transferAmount+1)/2) + approveAmount_1;
       let text = `approve,constructor,,${totalSupply},,,false\napprove,transfer,instance,accounts[${transferToAccountIndex}] ${transferAmount},0,,false\napprove,approve,instance,accounts[${approveAccountIndex}] ${approveAmount_1},${transferToAccountIndex},,false\napprove,approve,instance,accounts[${approveAccountIndex}] ${approveAmount_2},${transferToAccountIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
 
@@ -101,10 +104,13 @@ helper.range(transactionCounts).forEach(l => {
       let burnAmount_1 = helper.random(1, (transferAmount+1)/2);
       let burnAmount_2 = helper.random(1, (transferAmount+1)/2);
       let text = `burn,constructor,,${totalSupply},,,false\nburn,transfer,instance,accounts[${transferToAccountIndex}] ${transferAmount},,,false\nburn,burn,instance,accounts[${transferToAccountIndex}] ${burnAmount_1},,,false\nburn,burn,instance,accounts[${transferToAccountIndex}] ${burnAmount_2},,,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     })   
   }  
 
@@ -119,10 +125,13 @@ helper.range(transactionCounts).forEach(l => {
       let freezeAmount_1 = helper.random(1, (transferAmount+1)/2);
       let freezeAmount_2 = helper.random(1, (transferAmount+1)/2);
       let text = `freeze,constructor,,${totalSupply},,,false\nfreeze,transfer,instance,accounts[${transferToAccountIndex}] ${transferAmount},,,false\nfreeze,freeze,instance,${freezeAmount_1},${transferToAccountIndex},,false\nfreeze,freeze,instance,${freezeAmount_2},${transferToAccountIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     })  
   }
 
@@ -145,10 +154,13 @@ helper.range(transactionCounts).forEach(l => {
       let transferFinalAmount_1 = helper.random(1, (transferAmount+1)/2);
       let transferFinalAmount_2 = helper.random(1, (transferAmount+1)/2);
       let text = `transfer,constructor,,${totalSupply},,,false\ntransfer,transfer,instance,accounts[${transferToAccountIndex}] ${transferAmount},,,false\ntransfer,transfer,instance,accounts[${transferToFinalAccountIndex}] ${transferFinalAmount_1},${transferToAccountIndex},,false\ntransfer,transfer,instance,accounts[${transferToFinalAccountIndex}] ${transferFinalAmount_2},${transferToAccountIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     })  
   }  
 
@@ -174,10 +186,13 @@ helper.range(transactionCounts).forEach(l => {
       let transferFromAmount_1 = helper.random(1, (approveAmount+1)/2);
       let transferFromAmount_2 = helper.random(1, (approveAmount+1)/2);
       let text = `transferFrom,constructor,,${totalSupply},,,false\ntransferFrom,transfer,instance,accounts[${transferToAccountIndex}] ${transferAmount},,,false\ntransferFrom,approve,instance,accounts[${approveAccountIndex}] ${approveAmount},${transferToAccountIndex},,false\ntransferFrom,transferFrom,instance,accounts[${transferToAccountIndex}] accounts[${transferFromToIndex}] ${transferFromAmount_1},${approveAccountIndex},,false\ntransferFrom,transferFrom,instance,accounts[${transferToAccountIndex}] accounts[${transferFromToIndex}] ${transferFromAmount_2},${approveAccountIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     })  
 
   }
@@ -193,10 +208,13 @@ helper.range(transactionCounts).forEach(l => {
       let unfreezeAmount_1 = helper.random(1, (freezeAmount+1)/2);
       let unfreezeAmount_2 = helper.random(1, (freezeAmount+1)/2);
       let text = `unfreeze,constructor,,${totalSupply},,,false\nunfreeze,transfer,instance,accounts[${transferToAccountIndex}] ${transferAmount},,,false\nunfreeze,freeze,instance,${freezeAmount},${transferToAccountIndex},,false\nunfreeze,unfreeze,instance,${unfreezeAmount_1},${transferToAccountIndex},,false\nunfreeze,unfreeze,instance,${unfreezeAmount_2},${transferToAccountIndex},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     }) 
   }
   if(transactionName == 'withdrawEther') {
@@ -210,10 +228,13 @@ helper.range(transactionCounts).forEach(l => {
       let withdrawAmount_1 = helper.random(1, valueAmount);
       let withdrawAmount_2 = helper.random(1, valueAmount);
       let text = `withdrawEther,constructor,,${totalSupply},${owner},,false\nwithdrawEther,withdrawEther,instance,${withdrawAmount_1},${owner},web3.utils.toWei(${withdrawAmount_1} ether),false\nwithdrawEther,withdrawEther,instance,${withdrawAmount_2},${owner},web3.utils.toWei(${withdrawAmount_2} ether),true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     })   
   }  
 
@@ -228,10 +249,13 @@ helper.range(transactionCounts).forEach(l => {
       let mintAmount_2 = helper.random(lowerBoundInput, upperBoundInput+1);
       let mintAccountIndex = helper.random(0, deployAccountCount);
       let text = `mint,constructor,,${totalSupply},${owner},,false\nmint,mint,instance,accounts[${mintAccountIndex}] ${mintAmount_1},${owner},,false\nmint,mint,instance,accounts[${mintAccountIndex}] ${mintAmount_2},${owner},,true\n`;
-      fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
-        if (err) throw err;
-        console.log('File is created successfully.');
-      });
+      if(!fs.existsSync(path.join(transactionFolderPath, fileName))) {
+        console.log('generating new tracefiles ...');
+        fs.writeFileSync(path.join(transactionFolderPath, fileName), text, function (err) {
+          if (err) throw err;
+          console.log('File is created successfully.');
+        });
+      }
     })   
   }
 })
